+++ CHANGELOG.md
@@ -7,6 +7,7 @@ see examples:
   * `examples/param-decorator.php`
   * `examples/param-decorator2.php`
   * `examples/param-decorator-item.php`
+* Changed - cannot add pattern after unserialize `Awesomite\Chariot\Pattern\Patterns`
 * Changed - cannot add route after restoring router from cache
 ```php
 <?php